+++ Codigos/searches.cpp
@@ -3,7 +3,7 @@
 
 using namespace std;
 
-std::vector<int> binarySearch(int elem, vector<int> &vet, int n)
+std::vector<long long> binarySearch(int elem, vector<int> &vet, int n)
 {
     int low = 0;
     int comp = 0;
@@ -32,7 +32,7 @@ std::vector<int> binarySearch(int elem, vector<int> &vet, int n)
     return {-1,comp};
 }
 
-std::vector<int> linearSearch(int elem, vector<int> &vet, int n)
+std::vector<long long> linearSearch(int elem, vector<int> &vet, int n)
 {
 
     int comp =0;
